pyserial_to_arduino-main(testing)/pyserial_to_arduino-main/requirements.txt
altgraph==0.17.3
bidict==0.22.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.1.3
Flask-SocketIO==5.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pefile==2023.2.7
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyserial==3.5
python-engineio==4.3.3
python-socketio==5.7.1
pywin32-ctypes==0.2.0
requests==2.28.1
serial-tool==0.0.1
simple-websocket==0.7.0
urllib3==1.26.14
Werkzeug==2.0.1
winaudio==1.0.6
wincertstore==0.2
wsproto==1.2.0
